pyproject.toml
[project]
name="xumes"
version="0.0.41"
authors = [
  { name="Vincent Mastain", email="vincentmastain@gmail.com" },
]
description = "A game test tool."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "gymnasium>=0.28.1",
    "pygame>=2.4.0",
    "numpy>=1.23.5",
    "pyzmq>=25.1.0",
    "flask>=2.3.2",
    "sb3_contrib>=2.0.0a1",
    "stable_baselines3>=2.0.0a1",
    "requests>=2.30.0",
    "click>=8.1.3",
    "statsmodels>=0.14.0",
    "gherkin-official>=24.1.0",
    "multiprocess>=0.70.14",
    "sphinx-rtd-theme>=1.2.2"
]


[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"
[project.urls]
"Homepage" = "https://github.com/mastainvin/Xumes"
"Bug Tracker" = "https://github.com/mastainvin/Xumes/issues"

[project.scripts]
xumes = "xumes.core.xumes_cli:cli"
